Get rid of DirectX12 backend
This is currently taking up too much effort since
I effectively have to wrap every Vulkan/DirectX command and type into
some kind of abstraction, even when it doesn't add any value.
As this project's goal is to create a modern renderer,
and not to learn every single GPU API in existence, it
will be a lot easier to support only a single API.
Since DirectX 12 is poorly documented, and Vulkan is available
in one form or another on most platforms, its choice over DirectX
is pretty obvious.